+++ src/main/java/io/gdcc/spi/export/croissant/CroissantExporter.java
@@ -16,6 +16,8 @@ import jakarta.json.JsonReader;
 import jakarta.json.JsonValue;
 import jakarta.ws.rs.core.MediaType;
 
+import org.apache.commons.text.StringEscapeUtils;
+
 import java.io.OutputStream;
 import java.io.StringReader;
 import java.util.ArrayList;
@@ -144,9 +146,11 @@ public class CroissantExporter implements Exporter {
 
             JsonObject datasetORE = dataProvider.getDatasetORE();
             JsonObject describes = datasetORE.getJsonObject("ore:describes");
-            job.add("name", describes.getString("title"));
+            job.add("name", StringEscapeUtils.escapeHtml4(describes.getString("title")));
             job.add("url", describes.getJsonString("@id"));
             JsonObject datasetSchemaDotOrg = dataProvider.getDatasetSchemaDotOrg();
+            // We don't escape DatasetSchemaDotOrg fields like creator, description, etc. because
+            // they are already escaped.
             job.add("creator", datasetSchemaDotOrg.getJsonArray("creator"));
             job.add("description", datasetSchemaDotOrg.getJsonString("description"));
             job.add("keywords", datasetSchemaDotOrg.getJsonArray("keywords"));
@@ -205,9 +209,11 @@ public class CroissantExporter implements Exporter {
                  * in this? And would we duplicate all the cr:RecordSet entries (columns) with each
                  * additional format? Probably not as it would be the same.
                  */
-                String filename = fileDetails.getString("originalFileName", null);
+                String filename =
+                        StringEscapeUtils.escapeHtml4(
+                                fileDetails.getString("originalFileName", null));
                 if (filename == null) {
-                    filename = fileDetails.getString("filename");
+                    filename = StringEscapeUtils.escapeHtml4(fileDetails.getString("filename"));
                 }
                 String fileFormat = fileDetails.getString("originalFileFormat", null);
                 if (fileFormat == null) {
@@ -230,7 +236,8 @@ public class CroissantExporter implements Exporter {
                 String checksumType = checksum.getString("type").toLowerCase();
                 String checksumValue = checksum.getString("value");
                 String contentUrl = oreFiles.getJsonObject(fileCounter).getString("schema:sameAs");
-                String description = fileDetails.getString("description", "");
+                String description =
+                        StringEscapeUtils.escapeHtml4(fileDetails.getString("description", ""));
                 /**
                  * See https://github.com/mlcommons/croissant/issues/639 for discussion with the
                  * Croissant spec leads on what to put in
@@ -242,6 +249,7 @@ public class CroissantExporter implements Exporter {
                  *     https://github.com/IQSS/dataverse/issues/10523
                  */
                 String fileId = filename;
+                // We don't escape directory label because many characters aren't allowed anyway
                 String directoryLabel =
                         oreFiles.getJsonObject(fileCounter)
                                 .getString("dvcore:directoryLabel", null);
@@ -278,8 +286,11 @@ public class CroissantExporter implements Exporter {
                         JsonObject dataVariableObject = dataVariableValue.asJsonObject();
                         // TODO: should this be an integer?
                         Integer variableId = dataVariableObject.getInt("id");
-                        String variableName = dataVariableObject.getString("name");
-                        String variableDescription = dataVariableObject.getString("label", "");
+                        String variableName =
+                                StringEscapeUtils.escapeHtml4(dataVariableObject.getString("name"));
+                        String variableDescription =
+                                StringEscapeUtils.escapeHtml4(
+                                        dataVariableObject.getString("label", ""));
                         String variableFormatType =
                                 dataVariableObject.getString("variableFormatType");
                         String variableIntervalType =